deploy_temp/rembg-microservice/requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
rembg[cpu]>=2.0.62
python-multipart>=0.0.6
onnx>=1.15.0
onnxruntime>=1.17.0

# Optional: For GPU acceleration (uncomment if using CUDA)
# onnxruntime-gpu>=1.17.0
# torch>=2.1.0
# torchvision>=0.16.0

# Optional: For production deployment
# gunicorn>=21.2.0
# python-dotenv>=1.0.0
